requirements.txt
numpy
Pillow
pytesseract
pynput
